Fix timeout request type from int to float for accurate handling

+++ main.py
@@ -8,7 +8,7 @@ except ImportError:
     print("python-dotenv no está instalado, no se cargarán las variables de entorno, salvo que se usen en el .env.")
 
 DOMAINS = os.getenv("DOMAINS", "").split("|")
-TIMEOUT_REQUEST = int(os.getenv("TIMEOUT_REQUEST", 30))
+TIMEOUT_REQUEST = float(os.getenv("TIMEOUT_REQUEST", 30))
 RECEPTORES = os.getenv("RECEPTORES", "").split(";")
 
 def check_health(domain):
